requirements.txt
fastapi
uvicorn[standard]
aiosmtplib
python-dotenv
pydantic
pydantic-settings
email-validator
httpx
jinja2
slowapi
requests

asyncpg
sqlalchemy[asyncio]
alembic
clerk-backend-api
svix
pyjwt
